Save and load initial conditions, unit test to see if the saving and loading is correct

+++ include/Node.hh
@@ -15,12 +15,12 @@ public:
 
   void addToNetwork(Network* networkPtr);
   void assignToNetwork(Network* network);
-  void setValues(int* values);
+  void setValues(double* values);
 
   int getId(void) const;
   Network* getNetworkAssigned(void) const;
   std::vector<Network*> getNetworks(void) const;
-  void getValues(int* values);
+  void getValues(double* values);
 
   friend bool operator==(const Node& node1, const Node& node2);
 
@@ -28,6 +28,6 @@ private:
   int mNodeId;
   Network* mNetworkAssigned;
   std::vector<Network*> mNetworks;
-  int* mValues;
+  double* mValues;
 };
 #endif
